--- libs/ui/group-details/group-details.component.html
@@ -1,20 +1,0 @@
-<div *ngIf="selectedGroup() as group" class="m-2 p-2 border-2 rounded-md">
-  <h2 class="text-xl font-bold">Group Details</h2>
-  <p><strong>ID:</strong> {{ group.id }}</p>
-  <p><strong>Name:</strong> {{ group.name }}</p>
-  <p><strong>Description:</strong> {{ group.description }}</p>
-  <p><strong>Expanded:</strong> {{ group.expanded ? 'Yes' : 'No' }}</p>
-
-  <div *ngIf="group.subGroups.length > 0">
-    <h3 class="text-lg font-bold">Subgroups:</h3>
-    <ul>
-      <li *ngFor="let subGroup of group.subGroups">
-        {{ subGroup.name }}
-      </li>
-    </ul>
-  </div>
-</div>
-
-<div *ngIf="!selectedGroup()">
-  <p><strong>No group selected. Click a group to view details.</strong></p>
-</div>
