+++ libs/ui/recursive-group-list/src/components/group-list.component.html
@@ -1,0 +1,9 @@
+@if (selectedGroup) {
+    <clx-group-details [selectedGroup]="selectedGroup"></clx-group-details>
+}
+
+@if (groups) {
+    <div>
+        <clx-recursive-group-list [groups]="groups" (selectedGroupChange)="onSelectedGroupChange($event)"></clx-recursive-group-list>
+    </div>
+}
