+++ libs/ui/recursive-group-list/src/components/recursive-group-list/recursive-group-list.component.html
@@ -3,12 +3,12 @@
     <div class="m-1">
       <div
         (click)="onGroupClick(group)"
-        [ngClass]="{'expanded cursor-pointer transition-colors duration-300': group.isExpanded(), 'collapsed': !group.isExpanded(), 'hover: cursor-pointer': group.isGroup()}"
+        [ngClass]="{'expanded cursor-pointer transition-colors duration-300': isExpanded(group), 'collapsed': !isExpanded(group), 'hover: cursor-pointer': isGroup(group)}"
         class="group-node p-2.5 border-2 rounded-md flex items-center justify-start"
         (contextmenu)="onRightClick($event, group)">
 
-        @if (group.isGroup()) {
-          @if (group.isExpanded()) {
+        @if (isGroup(group)) {
+          @if (isExpanded(group)) {
             <span><i class="fa fa-chevron-up"></i></span>
           }
           @else {
@@ -19,7 +19,7 @@
         <span class="group-name ml-2.5 font-bold">{{ group.name }}</span>
       </div>
 
-      @if (group.isExpanded()) {
+      @if (isExpanded(group) && !!group.subGroups) {
         <div class="group-subgroups ml-5 p-1.25">
           <clx-recursive-group-list
             [groups]="group.subGroups"
@@ -28,16 +28,13 @@
         </div>
       }
     </div>
+
+    <clx-context-menu
+      [visible]="contextMenuVisible"
+      [position]="contextMenuPosition"
+      (rename)="renameGroup($event, group)"
+      (delete)="deleteGroup($event)"
+      (close)="contextMenuVisible = false">
+    </clx-context-menu>
   }
 }
-
-
-<clx-context-menu
-  [visible]="contextMenuVisible"
-  [position]="contextMenuPosition"
-  (rename)="renameGroup($event)"
-  (delete)="deleteGroup($event)"
-  (close)="contextMenuVisible = false">
-</clx-context-menu>
-
-
