+++ libs/ui/recursive-group-list/src/models/group-models.ts
@@ -1,8 +1,8 @@
-export interface IGroupData {
+export interface GroupModel {
   id: string;
   name: string;
   description: string;
   expanded: boolean;
   parentId?: string;
-  subGroups?: IGroupData[];
+  subGroups?: GroupModel[];
 }
