+++ src/app/app.component.html
@@ -1,12 +1,4 @@
 <!-- the following h1 can be removed upon implementation -->
 <h1 class="text-3xl font-bold pt-2 pb-2 underline mb-4 text-center">Recursive Group Examples</h1>
 
-@if (selectedGroup()) {
-  <clx-group-details [selectedGroup]="selectedGroup"></clx-group-details>
-}
-
-@if (groups()) {
-  <div>
-    <clx-recursive-group-list [groups]="groups()" (selectedGroupChange)="onSelectedGroupChange($event)"></clx-recursive-group-list>
-  </div>
-}
+<clx-group-list [groups]="groups" (selectedGroupChange)="onSelectedGroupChange($event)"></clx-group-list>
