--- src/models/group-methods.model.ts
@@ -1,9 +1,0 @@
-import { Group } from '.';
-
-export interface IGroupMethods {
-  toggleGroup(): void;
-  isGroup(): boolean;
-  isExpanded(): boolean;
-  selectedGroup(group: Group): Group | null;
-  rename(newName: string): void;
-}
